feat: URL ignore list
This allows people to exclude URLs from checking. Some URLs may only be
available in certain (company) networks, or may be rate limited, or may
behave just plain weirdly.

In my case, my POD pointed to https://developer.twitter.com/apps/ When you
go to that URL in a browser, you get redirected to a login page and from
there to your Twitter developer settings. In curl and LWP::UserAgent you get
a HTTP 405 ("method not allowed" seems odd here, I'd expect 3xx "redirect"
or a 401 "not authorized"). This 405 is treated as an error cause it's not a
200 "ok", which makes sense from Test::Pod::No404s' point of view, and hence
the test fails. Said Twitter URL is also rate-limited and hitting it too
often gets you (temporarily) blocked.

With this patch I can add the bad URL to my .no404s-ignore file and have
Test::Pod::No404s ignore that URL. This makes my tests pass and Twitter
won't block me if I run my tests ten times in a row, just because I can.

There may be better ways for finding that ignore file (e.g., look in
multiple locations), or the module could be extended to allow passing
configuration arguments somehow. I wanted to keep the interface the same and
the changes simple, hence this approach.

This patch does not include version number changes. I'd normally bump it to
0.3 as it's compatible with old code, but now new code cannot check the
version number to know whether the ignore files is supported. But I don't
know how this module is versioned and distributed.